requirements.txt
PySide6>=6.8.0
pynput>=1.7.7
pyperclip>=1.9.0
google-genai>=1.33.0
python-dotenv>=1.0.1